backend/requirements.txt
Django
django-cors-headers
Pillow
python-decouple
PyJWT
